feat: add configuration for hotwire native

+++ app/controllers/concerns/authentication.rb
@@ -39,6 +39,8 @@ module Authentication
     end
 
     def after_authentication_url
+      return refresh_app_path if hotwire_native_app?
+
       session.delete(:return_to_after_authenticating) || root_url
     end
